Going on fire with Barcode Scanner
+ Added a feature which allows to look for product info by simply
scanning its barcode:
If the product exists in the DB, it shows all the related infos,
otherwise you can add the product to the DB by clicking the related
button.
However, at the current state you can only view the product info. Next
step: show all the product prices.

+ In the previous commit I added a button on the main activity which is
shown as "BACKUPPONE NAZIONALE". It's just a temporary thingy that
should allow to backup the DB on a sd-card in the future. However, I
wasn't able to complete it, since it works only on old Android APIs. In
fact, since Android 5.0 things with the storage permissions went a
little bit messy. So, a lot of patience required.